+++ packages/sdk-core/CHANGELOG.md
@@ -9,6 +9,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - `increaseAllowance` and `decreaseAllowance` functions to SuperToken class
 - Batch call support for `increaseAllowance` and `decreaseAllowance`
 - `increaseFlowRateAllowance` and `decreaseFlowRateAllowance` functions to CFAv1 class
+- Add support for `activeOutgoingStreamCount`, `activeIncomingStreamCount`, `inactiveOutgoingStreamCount`, `inactiveIncomingStreamCount` properties in query for `AccountTokenSnapshot` entity
+- Add support for `rewardAddress`, `liquidationPeriod`, `patricianPeriod`, and `minimumDeposit` properties in query for `Token` entity
+- Add support for `userData` property in query for `Stream` entity
+
+### Changed
+- `.exec` or `.execForward` will throw an error if the operations array is empty
 
 ## [0.6.2] - 2023-02-07
 ### Added